requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
PyMySQL==1.1.1
python-dotenv==1.0.1
timeago
cryptography